+++ phidl/routing.py
@@ -13,7 +13,7 @@ def _arc(radius = 10, width = 0.5, theta = 45, start_angle = 0, angle_resolution
     outer_radius = radius+width/2
     angle1 = (start_angle)*pi/180
     angle2 = (start_angle + theta)*pi/180
-    t = np.linspace(angle1, angle2, np.ceil(abs(theta)/angle_resolution))
+    t = np.linspace(angle1, angle2, int(np.ceil(abs(theta)/angle_resolution)))
     inner_points_x = (inner_radius*cos(t)).tolist()
     inner_points_y = (inner_radius*sin(t)).tolist()
     outer_points_x = (outer_radius*cos(t)).tolist()
